requirements.txt
requests>=2.25.0
termcolor>=1.1.0
flask>=2.3.0
weave>=0.50.0
daytona-sdk>=0.110.0
python-dotenv>=0.19.0
